backend/requirements.txt
Django
djangorestframework
django-cors-headers
psycopg2-binary
Pillow